+++ app/components/views/Notifications.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const Notifications = () => (
+  <div>
+    <h1>Send Push Notifications</h1>
+    <p>COMING SOON!</p>
+  </div>
+)
+
+export default Notifications
